split Evaluator/Impl into core/src
