+++ build.mill
@@ -491,7 +491,8 @@ trait MillScalaModule extends ScalaModule with MillJavaModule with ScalafixModul
         "-Wconf:msg=import scala.language.implicitConversions:silent",
         "-Wconf:msg=IterableOnceExtensionMethods:silent",
         "-Wconf:msg=is deprecated:silent",
-        "-Wconf:msg=cannot be checked at runtime:silent"
+        "-Wconf:msg=cannot be checked at runtime:silent",
+        "-Wconf:msg=unused:silent"
         // "-no-indent",
         // "-Wvalue-discard",
         // "-Wshadow:all",
