+++ core/eval/src/mill/eval/Cached.scala
@@ -1,0 +1,7 @@
+package mill.eval
+
+case class Cached(value: ujson.Value, valueHash: Int, inputsHash: Int)
+
+object Cached {
+  implicit val rw: upickle.default.ReadWriter[Cached] = upickle.default.macroRW
+}
