+++ core/eval/src/mill/eval/TaskResult.scala
@@ -1,0 +1,9 @@
+package mill.eval
+import mill.api.Result
+
+case class TaskResult[T](result: Result[T], recalc: () => Result[T]) {
+  def map[V](f: T => V): TaskResult[V] = TaskResult[V](
+    result.map(f),
+    () => recalc().map(f)
+  )
+}
