+++ testkit/src/mill/testkit/IntegrationTester.scala
@@ -2,7 +2,7 @@ package mill.testkit
 
 import mill.main.client.EnvVars.MILL_TEST_SUITE
 import mill.define.Segments
-import mill.eval.Evaluator
+import mill.eval.{Evaluator, Cached}
 import mill.main.client.OutFiles
 import mill.resolve.SelectMode
 import ujson.Value
@@ -121,7 +121,7 @@ object IntegrationTester {
        * Returns the `.json` metadata file contents parsed into a [[Evaluator.Cached]]
        * object, containing both the value as JSON and the associated metadata (e.g. hashes)
        */
-      def cached: Evaluator.Cached = upickle.default.read[Evaluator.Cached](text)
+      def cached: Cached = upickle.default.read[Cached](text)
 
       /**
        * Returns the value as JSON
